pyproject.toml
[project]
name = "toolchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Andrew Wason", email = "rectalogic@rectalogic.com" }]
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "pydantic-ai>=0.0.44",
    "pydantic-mcp>=0.1.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]

[project.scripts]
toolchat = "toolchat.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.11.2"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I",
    # flake8-no-pep420
    "INP",
    # pyupgrade
    "UP",
]
